Renamed kunta rekry resource containers
